getYValueFor ahora devuelve float

+++ AlumnoEjemplos/RestrictedGL/Interfaces/ITerrainCollision.cs
@@ -3,7 +3,7 @@
 namespace AlumnoEjemplos.RestrictedGL.Interfaces {
     public interface ITerrainCollision {
 
-        int getYValueFor(float x, float z);
+        float getYValueFor(float x, float z);
 
         void deform(float x, float z, float radius, int power);
